+++ allocator.h
@@ -25,13 +25,14 @@
 #define __ALLOCATOR_H__
 
 #include <sys/types.h>
+#include <inttypes.h>
 
 void slab_init();
 void slab_cleanup(int quiet);
-void *slab_alloc(void *p, size_t size);
-void *slab_calloc(void *p, size_t items, size_t size);
+void *slab_alloc(void *p, uint64_t size);
+void *slab_calloc(void *p, uint64_t items, uint64_t size);
 void slab_free(void *p, void *address);
-int slab_cache_add(size_t size);
+int slab_cache_add(uint64_t size);
 
 #endif
